requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langgraph>=0.0.66
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.12
pydantic>=2.5.2
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
aiofiles>=23.2.1
pandas>=2.1.4
numpy>=1.26.0
scikit-learn>=1.3.2
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.25
alembic>=1.13.1
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0 